feat: updates to actor sheet colors (#1188)
This update allows the user to set the primary and secondary colors for the default Twodsix theme (not the Foundry default theme).  There are two settings in "Display Settings": Default Color (main color) and Highlight Color.   Clicking on the color sample brings up the color selector popup.  Note: To return to the default configuration, delete the hex codes in the boxes and save settings.

This update also brings improvements to ES (@ForjaSalvaje) and FR(@ArtheniusRPG) localizations