feat: add token button to request roll (#1303)
This feature adds the ability for the GM to request a specific roll/check through chat.  This is done by the GM, first, selecting tokens from the active scene to make a roll, then clicking a new button in the token select interface, and choosing roll parameters through a pop-up.  Players can make these requested rolls though a special chat message. Detailed instructions on how to use this feature are on the wiki: https://github.com/xdy/twodsix-foundryvtt/wiki/Requesting-a-Roll-from-Selected-Tokens
Thanks to @romanK for the suggestion and input from @DaishoChikara, @Debinani, and @Simyonovich.
Also included are some ES translation updates from @ForjaSalvaje and a fix to allow Augments to have consumables used during rolls - per suggestion of @RadioFreeWABA