feat: add non-stackable armor (#1458)
This update adds a checkbox to the armor sheet to designate it as non-stackable (cannot be worn with other armor).  When such armor is equipped, a warning will appear when trying to equip additional armor (although the user is not prevented from doing so). Non-stackable armor that is equipped will have a red "layers" icon after its name on the actor sheet.  Thanks to @romanK for the suggestion.

Also, this update includes localization updates for ES from @ForjaSalvaje and a very extensive update for DE from @Mithnar.