feat: add setting to ignore effect for damage rolls (#1496)
This release adds a new rules option (Advanced Ruleset Settings -> Damage -> Add effect when rolling damage) to ignore the skill roll effect when calculating damage (useful for Classic Traveller).  The default is true - which includes the effect in damage rolls (legacy behavior).  Thanks to @Hideo Kojima for the suggestion.

Also,  the TwodsixDiceRolls and TwodsixRollSettings are now exposed to user macros (e.g, with game.twodsix.TwodsixDiceRolls)

And some fixes:
* autocomplete now works correctly for characteristics
* improvements to the automatic status effects code to reduce calls to updateActor